added the structs for the elf

+++ matrix_kernel/src/start.cpp
@@ -4,9 +4,9 @@
 
 extern "C" void _start(mbi::boot_info info)
 {
-    for (int x = 0; x < info.m_buffer.width(); x++)
+    for (uint64 x = 0; x < info.m_buffer.width(); x++)
     {
-        for (int y = 0; y < info.m_buffer.height(); y++)
+        for (uint64 y = 0; y < info.m_buffer.height(); y++)
         {
             info.m_buffer.set_pixel(x, y, {0, 255, 0, 0});
         }
